+++ NEWS.md
@@ -12,6 +12,7 @@
 
 * `eval_code` uses `evaluate::evaluate` and stores returned outputs in the code's attribute.
 * Refactor `eval_code` method signature to allow for more flexibility when extending the `eval_code`/`within` functions.
+* `get_var(qenv, ...)` and `join(qenv, ...)` were hard deprecated.
 
 # teal.code 0.6.1